+++ src/main/java/com/example/postitback/repositories/PostRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PostRepository extends JpaRepository<Posts, Long> {
-    List<Posts> findAllByUserId(Long id);
+public interface PostRepository extends JpaRepository<Posts, Integer> {
+    List<Posts> findAllByUserId(Integer id);
 }
